Refactor: 어드민 Qna 리팩토링 -> DTO 타입 일부 변경, 코드 스타일링, QnaService @Transactional 추가

+++ src/main/java/com/tbp/honeyjar/admin/dto/qna/QnaCorrectionRequestDto.java
@@ -7,9 +7,9 @@ import java.time.LocalDateTime;
 @Data
 public class QnaCorrectionRequestDto {
 
-    private long id;
+    private Long id;
     private String title;
     private String post;
-    private int categoryId;
+    private Long categoryId;
     private LocalDateTime updatedAt;
 }
